# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Contract Manage API",
        "description": "API documentation for Contract Management System",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Contract Manage API Server"
        }
    ],
    "paths": {
...
        "/api/contracts": {
            "get": {
                "tags": [
                    "Contracts"
                ],
                "summary": "Get list of contracts",
                "operationId": "1ab3a5b4a5e2aafb0f49ce993ee8a56a",
                "parameters": [
                    {
                        "name": "view",
                        "in": "query",
                        "description": "Filter by view (dashboard, contracts, mine, pending, etc.)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of contracts"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Contracts"
                ],
                "summary": "Create a new contract",
                "operationId": "ec9937d8f85a0ef70236302888ea421e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "contract_type_id": {
                                        "type": "string"
                                    },
                                    "submission_type_id": {
                                        "type": "string"
                                    },
                                    "vendor_id": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contract created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}